+++ ghostwriter/rolodex/forms_project.py
@@ -2,8 +2,10 @@
 
 # Django Imports
 from django import forms
+from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.forms.models import BaseInlineFormSet, inlineformset_factory
+from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 
 # 3rd Party Libraries
@@ -25,18 +27,22 @@ from crispy_forms.layout import (
 from ghostwriter.modules.custom_layout_object import CustomTab, Formset, SwitchToggle
 
 from .models import (
+    Deconfliction,
     Project,
     ProjectAssignment,
     ProjectNote,
     ProjectObjective,
     ProjectScope,
     ProjectTarget,
+    WhiteCard,
 )
 
 # Number of "extra" formsets created by default
 # Higher numbers can increase page load times with WYSIWYG editors
 EXTRAS = 0
 
+
+# Custom inline formsets for nested forms
 
 class BaseProjectObjectiveInlineFormSet(BaseInlineFormSet):
     """
@@ -334,6 +340,54 @@ class BaseProjectTargetInlineFormSet(BaseInlineFormSet):
                             ),
                         )
 
+
+class BaseWhiteCardInlineFormSet(BaseInlineFormSet):
+    """
+    BaseInlineFormset template for :model:`rolodex.WhiteCard` that adds validation
+    for this model.
+    """
+
+    def clean(self):
+        super().clean()
+        if any(self.errors):  # pragma: no cover
+            return
+        for form in self.forms:
+            if form.cleaned_data:
+                # Only validate if the form is NOT marked for deletion
+                if form.cleaned_data["DELETE"] is False:
+                    title = form.cleaned_data["title"]
+                    issued = form.cleaned_data["issued"]
+
+                    # Check that all objective have a deadline and status
+                    if title and not issued:
+                        form.add_error(
+                            "issued",
+                            ValidationError(
+                                _("Your white card still needs an issued date and time"),
+                                code="incomplete",
+                            ),
+                        )
+                    elif issued and not title:
+                        form.add_error(
+                            "title",
+                            ValidationError(
+                                _("Your white card still needs a title"),
+                                code="incomplete",
+                            ),
+                        )
+                    # Raise an error if dates are out of bounds
+                    # We only check if ``issued`` is after the project's end date because white cards can be issued prior to execution
+                    if self.instance.start_date and issued:
+                        if issued.date() > self.instance.end_date:
+                            form.add_error(
+                                "issued",
+                                ValidationError(
+                                    _("Your selected date is after the project end date"),
+                                    code="invalid_datetime",
+                                ),
+                            )
+
+# Forms used with the inline formsets
 
 class ProjectAssignmentForm(forms.ModelForm):
     """
@@ -750,6 +804,101 @@ class ProjectTargetForm(forms.ModelForm):
         )
 
 
+class WhiteCardForm(forms.ModelForm):
+    """
+    Save an individual :model:`rolodex.WhiteCard` associated with an individual
+    :model:`rolodex.Project`.
+    """
+
+    class Meta:
+        model = WhiteCard
+        exclude = ("project",)
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        for field in self.fields:
+            self.fields[field].widget.attrs["autocomplete"] = "off"
+        self.fields["issued"].widget.input_type = "datetime-local"
+        self.fields["description"].widget.attrs["rows"] = 5
+        self.fields["description"].widget.attrs[
+            "placeholder"
+        ] = "Additional information about the white card, the reason for it, limitations, how it affects the assessment, etc..."
+        self.fields["title"].widget.attrs["placeholder"] = "Brief and Descriptive Headline"
+        self.helper = FormHelper()
+        self.helper.form_show_errors = False
+        # Disable the <form> tags because this will be inside an instance of `ProjectForm()`
+        self.helper.form_tag = False
+        # Disable CSRF so `csrfmiddlewaretoken` is not rendered multiple times
+        self.helper.disable_csrf = True
+        # Hide the field labels from the model
+        self.helper.form_show_labels = False
+        # Layout the form for Bootstrap
+        self.helper.layout = Layout(
+            # Wrap form in a div so Django renders form instances in their own element
+            Div(
+                # These Bootstrap alerts begin hidden and function as undo buttons for deleted forms
+                Alert(
+                    content=(
+                        """
+                        <strong>White Card Deleted!</strong>
+                        Deletion will be permanent once the form is submitted. Click this alert to undo.
+                        """
+                    ),
+                    css_class="alert alert-danger show formset-undo-button",
+                    style="display:none; cursor:pointer;",
+                    template="alert.html",
+                    block=False,
+                    dismiss=False,
+                ),
+                Div(
+                    HTML(
+                        """
+                        <h6>White Card #<span class="counter">{{ forloop.counter }}</span></h6>
+                        <hr>
+                        """
+                    ),
+                    Row(
+                        Column("title", css_class="col-md-6"),
+                        Column(
+                            FieldWithButtons(
+                                Field("issued", step=1,),
+                                HTML(
+                                    """
+                                    <button
+                                        class="btn btn-secondary"
+                                        type="button"
+                                        onclick="setNow($(this).closest('div').find('input'))"
+                                    >
+                                    Now
+                                    </button>
+                                    """
+                                ),
+                            ),
+                            css_class="col-md-6"
+                        ),
+                    ),
+                    "description",
+                    Row(
+                        Column(
+                            Button(
+                                "formset-del-button",
+                                "Delete White Card",
+                                css_class="btn-sm btn-danger formset-del-button",
+                            ),
+                            css_class="form-group col-md-4 offset-md-4",
+                        ),
+                        Column(
+                            Field("DELETE", style="display: none;"),
+                            css_class="form-group col-md-4 text-center",
+                        ),
+                        css_class="form-row",
+                    ),
+                    css_class="formset",
+                ),
+                css_class="formset-container",
+            )
+        )
+
 # Create the `inlineformset_factory()` objects for `ProjectForm()`
 
 ProjectAssignmentFormSet = inlineformset_factory(
@@ -789,6 +938,15 @@ ProjectTargetFormSet = inlineformset_factory(
     can_delete=True,
 )
 
+WhiteCardFormSet = inlineformset_factory(
+    Project,
+    WhiteCard,
+    form=WhiteCardForm,
+    formset=BaseWhiteCardInlineFormSet,
+    extra=EXTRAS,
+    can_delete=True,
+)
+
 
 class ProjectForm(forms.ModelForm):
     """
@@ -818,16 +976,13 @@ class ProjectForm(forms.ModelForm):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.fields["start_date"].widget.attrs["autocomplete"] = "off"
-        self.fields["start_date"].widget.attrs["autocomplete"] = "off"
+        for field in self.fields:
+            self.fields[field].widget.attrs["autocomplete"] = "off"
         self.fields["start_date"].widget.input_type = "date"
-        self.fields["end_date"].widget.attrs["autocomplete"] = "off"
-        self.fields["end_date"].widget.attrs["autocomplete"] = "off"
         self.fields["end_date"].widget.input_type = "date"
         self.fields["start_time"].widget.input_type = "time"
         self.fields["end_time"].widget.input_type = "time"
         self.fields["slack_channel"].widget.attrs["placeholder"] = "#slack-channel"
-        self.fields["slack_channel"].widget.attrs["autocomplete"] = "off"
         self.fields["note"].widget.attrs["placeholder"] = "Description of the Project"
         # Hide labels for specific fields because ``form_show_labels`` takes priority
         self.fields["start_date"].label = False
@@ -981,6 +1136,27 @@ class ProjectForm(forms.ModelForm):
                     link_css_class="tab-icon list-icon",
                     css_id="targets",
                 ),
+                CustomTab(
+                    "White Cards",
+                    HTML(
+                        """
+                        <p class="form-spacer"></p>
+                        """
+                    ),
+                    Formset("whitecards", object_context_name="White Card"),
+                    Button(
+                        "add-whitecard",
+                        "Add White Card",
+                        css_class="btn-block btn-secondary formset-add-card",
+                    ),
+                    HTML(
+                        """
+                        <p class="form-spacer"></p>
+                        """
+                    ),
+                    link_css_class="tab-icon whitecard-icon",
+                    css_id="whitecards",
+                ),
                 template="tab.html",
                 css_class="nav-justified",
             ),
@@ -1055,3 +1231,106 @@ class ProjectNoteForm(forms.ModelForm):
                 code="required",
             )
         return note
+
+
+class DeconflictionForm(forms.ModelForm):
+    """
+    Save an individual :model:`rolodex.Deconfliction` associated with an individual
+    :model:`rolodex.Project`.
+    """
+
+    class Meta:
+        model = Deconfliction
+        exclude = ("created_at", "project",)
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        for field in self.fields:
+            self.fields[field].widget.attrs["autocomplete"] = "off"
+        self.fields["report_timestamp"].widget.input_type = "datetime-local"
+        self.fields["alert_timestamp"].widget.input_type = "datetime-local"
+        self.fields["response_timestamp"].widget.input_type = "datetime-local"
+        self.fields["report_timestamp"].label = "Date & Time of Report"
+        self.fields["alert_timestamp"].label = "Date & Time the Alert Triggered"
+        self.fields["response_timestamp"].label = "Date & Time of Your Response"
+        self.fields["title"].label = ""
+        self.fields["status"].label = ""
+        self.fields["alert_source"].label = ""
+        self.fields["description"].label = ""
+        self.fields["description"].widget.attrs["rows"] = 5
+        self.fields["description"].widget.attrs[
+            "placeholder"
+        ] = "Additional information about the alert, source, related activity..."
+        self.fields["title"].widget.attrs["placeholder"] = "Brief and Descriptive Title"
+        self.fields["alert_source"].widget.attrs["placeholder"] = "Source of the Alert – e.g, EDR"
+        self.fields["report_timestamp"].initial = timezone.now()
+        self.helper = FormHelper()
+        self.helper.form_show_errors = False
+        # Layout the form for Bootstrap
+        self.helper.layout = Layout(
+            Row(
+                Column("title", css_class="form-group col-12 mb-0"),
+                css_class="form-group",
+            ),
+            Row(
+                Column("status", css_class="form-group col-6 mb-0"),
+                Column("alert_source", css_class="form-group col-6 mb-0"),
+                css_class="form-group",
+            ),
+            HTML(
+                f"""
+                <p>You can update these timestamps as you get more information. Use the server's time zone ({settings.TIME_ZONE}).
+                """
+            ),
+            Row(
+                Column(Field("alert_timestamp", step=1), css_class="form-group col-4 mb-0"),
+                Column(Field("report_timestamp", step=1), css_class="form-group col-4 mb-0"),
+                Column(Field("response_timestamp", step=1), css_class="form-group col-4 mb-0"),
+                css_class="form-group",
+            ),
+            Row(
+                Column("description", css_class="form-group col-12 mb-0"),
+                css_class="form-group",
+            ),
+            ButtonHolder(
+                Submit("submit_btn", "Submit", css_class="btn btn-primary col-md-4"),
+                HTML(
+                    """
+                    <button onclick="window.location.href='{{ cancel_link }}'" class="btn btn-outline-secondary col-md-4" type="button">Cancel</button>
+                    """
+                ),
+            ),
+        )
+
+    def clean(self):
+        alert_timestamp = None
+        report_timestamp = None
+        response_timestamp = None
+
+        cleaned_data = super().clean()
+        if "alert_timestamp" in cleaned_data:
+            alert_timestamp = cleaned_data["alert_timestamp"]
+        if "report_timestamp" in cleaned_data:
+            report_timestamp = cleaned_data["report_timestamp"]
+        if "response_timestamp" in cleaned_data:
+            response_timestamp = cleaned_data["response_timestamp"]
+
+        if response_timestamp and report_timestamp:
+            if response_timestamp < report_timestamp:
+                self.add_error(
+                    "response_timestamp",
+                    ValidationError(
+                        _("The response timestamp cannot be before the report timestamp"),
+                        code="invalid_datetime",
+                    ),
+                )
+
+        if report_timestamp and alert_timestamp:
+            if report_timestamp < alert_timestamp:
+                self.add_error(
+                    "report_timestamp",
+                    ValidationError(
+                        _("The report timestamp cannot be before the alert timestamp"),
+                        code="invalid_datetime",
+                    ),
+                )